Add possibility to COPY diff-mentioned files in
When the image lacks both the `git` and `patch` binary, the standard way
of patching the image does not work. One possibility would be to use
`--run-before` to install things (and `--run-after` to remove them), but
that can be tedious and take long. Instead, we now offer a new way: by
passing `--use-copy-to-apply` the user can instruct `docker-image-patch`
to simply copy the whole files mentioned in the diff over into the
image. While there are still situations where this might not be enough,
e.g. multiple patches from multiple directories, this does work in the
basic use-case.